% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypted_html_document.R
\name{encrypt_html_file}
\alias{encrypt_html_file}
\title{Encrypt an html file}
\usage{
encrypt_html_file(path, output_path = paste0(path, ".enc.html"),
  key = sodium::random(32L), message_key = TRUE,
  write_key_file = FALSE,
  output_template_path = system.file("html-template.html", package =
  "encryptedRmd"))
}
\arguments{
\item{path}{the file you want to encrypt}

\item{output_path}{optional, the output path}

\item{key}{optional, the encryption key}

\item{message_key}{optional, print the encryption key to the console}

\item{write_key_file}{optional, write a key file in the same directory}

\item{output_template_path}{a path to the output template. The output template needs have the same html form elements (same ids) and the same placeholders as the default template. Everything else can be customized.}
}
\value{
The key used to encrypt the file in hex encoding as an invisible character vector.
}
\description{
Encrypt an html file
}
