% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypted_html_document.R
\name{encrypt_html_file}
\alias{encrypt_html_file}
\title{Encrypt an html file}
\usage{
encrypt_html_file(path, output_path = paste0(path, ".enc.html"),
  key = sodium::random(32L), message_key = TRUE,
  write_key_file = FALSE)
}
\arguments{
\item{path}{the file you want to encrypt}

\item{output_path}{optional, the output path}

\item{key}{optional, the encryption key}

\item{message_key}{optional, print the encryption key to the console}

\item{write_key_file}{optional, write a key file in the same directory}
}
\description{
Encrypt an html file
}
