% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypted_html_document.R
\name{encrypted_html_document}
\alias{encrypted_html_document}
\title{Convert to an encrypted HTML document}
\usage{
encrypted_html_document(...)
}
\arguments{
\item{...}{all parameters are passed to rmarkdown::html_document}
}
\description{
In addition to a standard html file the function also creates an encrypted version
together with the key as two separate files.
}
