% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypted_html_document.R
\name{encrypted_html_document}
\alias{encrypted_html_document}
\title{Create an encrypted HTML document}
\usage{
encrypted_html_document(...)
}
\arguments{
\item{...}{all parameters are passed to \code{rmarkdown::html_document}}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown]{render}}
}
\description{
In addition to a standard html file this \code{rmarkdown} format also creates
an encrypted version together with the key as two separate files.
}
\details{
Two files are created:
\describe{
 \item{filename.enc.html}{This is the password protected file.}
 \item{filename.enc.html.key}{This file contains the key with which the report was encrypted.}
}


Warning: You are using this at your own risk. Make sure your encryption key is
strong enough. For serious use cases, please also review the code of the functions.
Any feedback is appreciated. This is an early package version.
Please only share the key file with trusted parties
}
\seealso{
\code{\link{encrypt_html_file}} for more information on the encryption.
}
